frontend/requirements.txt
Flask==2.3.3
requests==2.31.0
jinja2==3.1.3
gunicorn==21.2.0
Flask-Session==0.4.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.0.0